Refactored and integrated ls and date and added tests to the new feature
