+++ src/uu/ls/Cargo.toml
@@ -23,7 +23,6 @@ chrono-tz = { workspace = true }
 clap = { workspace = true, features = ["env"] }
 glob = { workspace = true }
 hostname = { workspace = true }
-iana-time-zone = { workspace = true }
 lscolors = { workspace = true }
 number_prefix = { workspace = true }
 once_cell = { workspace = true }
@@ -31,6 +30,7 @@ selinux = { workspace = true, optional = true }
 terminal_size = { workspace = true }
 uucore = { workspace = true, features = [
   "colors",
+  "custom-tz-fmt",
   "entries",
   "format",
   "fs",
