+++ src/uucore/src/lib/features.rs
@@ -12,6 +12,8 @@ pub mod buf_copy;
 pub mod checksum;
 #[cfg(feature = "colors")]
 pub mod colors;
+#[cfg(feature = "custom-tz-fmt")]
+pub mod custom_tz_fmt;
 #[cfg(feature = "encoding")]
 pub mod encoding;
 #[cfg(feature = "format")]
